% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdf.R
\name{ecdf_ci}
\alias{ecdf_ci}
\title{Calculates confidence interval for eCDF}
\usage{
ecdf_ci(x, breaks = seq(-100, 100, 1), conf_level = 0.95)
}
\arguments{
\item{x}{Numeric vector of observations}

\item{breaks}{Values to calculate eCDF difference}

\item{conf_level}{Confidence level}
}
\value{
A data frame with the estimated cumulative distribution and
  confidence intervals for a range of values.
}
\description{
Returns a data frame of the estimated cumulative distribution and lower and
upper confidence intervals.
}
\details{
Given a vector \eqn{x_1 \sim F_1} this function calculates confidence
intervals for \eqn{F_1(x)} for \eqn{x} defined by \code{breaks}. Confidence
intervals are derived using \link[DescTools]{BinomCI} with the exact
Clopper-Pearson method for binomial confidence intervals.
}
\examples{
x1 <- sample(-100:100, 25)
ecdf_ci(x1)

}
