% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdf.R
\name{ecdf_diff}
\alias{ecdf_diff}
\title{Calculates difference between eCDFs}
\usage{
ecdf_diff(x1, x2, breaks = seq(-100, 100, 1), conf_level = 0.95)
}
\arguments{
\item{x1, x2}{Numeric vector of observations}

\item{breaks}{Values to calculate eCDF difference}

\item{conf_level}{Confidence level}
}
\value{
A data frame of the value, estimated difference and confidence
limits.
}
\description{
Returns a data frame of differences in the empirical distribution functions.
}
\details{
Given vectors \eqn{x_1 \sim F_1} and \eqn{x_2 \sim F_2}, this function
calculates \eqn{F_1(x) - F_2(x)} for \eqn{x} defined by \code{breaks}.
Confidence intervals are derived using \link[DescTools]{BinomDiffCI} with
the Mee (1984) and Farrington-Manning (1990) method for difference in
binomial proportions confidence interval.
}
\examples{
x1 <- sample(-100:100, 25)
x2 <- sample(-100:100, 20)
ecdf_diff(x1, x2)

}
