% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_waterfall}
\alias{plot_waterfall}
\title{Plots waterfall bar plot}
\usage{
plot_waterfall(x, title = "")
}
\arguments{
\item{x}{Vector of best tumor size changes.}

\item{title}{Title of the plot.}
}
\value{
A ggplot bar plot.
}
\description{
\code{plot_waterfall} returns a ggplot object which is the typical waterfall bar
plot with patients' best tumor size change sorted from largest to smallest.
}
\examples{
x <- sample(-100:100, 25)
plot_waterfall(x)

}
