% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_waterfall_curve}
\alias{plot_waterfall_curve}
\title{Plots waterfall curve}
\usage{
plot_waterfall_curve(
  x1,
  x2 = NULL,
  name1 = "Treatment",
  name2 = "Control",
  hline = 0
)
}
\arguments{
\item{x1}{Vector of best tumor size changes for treatment 1.}

\item{x2}{Vector of best tumor size changes for treatment 2.}

\item{name1}{Name for treatment 1.}

\item{name2}{Name for treatment 2.}

\item{hline}{Y-coordinate for dotted line.}
}
\value{
A ggplot step plot.
}
\description{
\code{plot_waterfall_curve} returns a ggplot object which presents best tumor
size change as a step function rather than the typical waterfall bar plot.
}
\examples{
x1 <- sample(-100:100, 25)
x2 <- sample(-100:100, 20)
plot_waterfall_curve(x1, x2)

}
