% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_waterfall_diff}
\alias{plot_waterfall_diff}
\title{Plots rate difference between waterfall curves}
\usage{
plot_waterfall_diff(
  x1,
  x2,
  name1 = "Treatment",
  name2 = "Control",
  conf_level = 0.95,
  hline = 0
)
}
\arguments{
\item{x1}{Vector of best tumor size changes for treatment 1}

\item{x2}{Vector of best tumor size changes for treatment 2}

\item{name1}{Name for treatment 1}

\item{name2}{Name for treatment 2}

\item{conf_level}{Confidence level}

\item{hline}{Y-coordinate for dotted line}
}
\value{
A ggplot step plot.
}
\description{
\code{plot_waterfall_diff} returns a ggplot object which is the rate
difference of two waterfall curves across the range of best tumor size
change.
}
\examples{
x1 <- sample(-100:100, 25)
x2 <- sample(-100:100, 20)
plot_waterfall_diff(x1, x2)

}
