% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_waterfall_curve}
\alias{plot_waterfall_curve}
\title{Plots waterfall curve}
\usage{
plot_waterfall_curve(
  x1,
  x2 = NULL,
  name1 = "Treatment",
  name2 = "Control",
  plot_ci1 = FALSE,
  plot_ci2 = FALSE,
  hline = 0,
  include_diff = TRUE,
  conf_level = 0.95
)
}
\arguments{
\item{x1}{Vector of best tumor size changes for treatment 1}

\item{x2}{Vector of best tumor size changes for treatment 2}

\item{name1}{Name for treatment 1}

\item{name2}{Name for treatment 2}

\item{plot_ci1}{Whether to plot confidence interval for x1. Default is FALSE.}

\item{plot_ci2}{Whether to plot confidence interval for x2. Default if FALSE.}

\item{hline}{Y-coordinate for dotted line}

\item{include_diff}{Whether to include difference in proportions plot}

\item{conf_level}{Confidence level for difference in proportions plot}
}
\value{
A ggplot step plot.
}
\description{
\code{plot_waterfall_curve} returns a ggplot object which presents best tumor
size change as a step function rather than the typical waterfall bar plot.
}
\examples{
x1 <- sample(-100:100, 25)
x2 <- sample(-100:100, 20)
plot_waterfall_curve(x1, x2)

}
